+++ src/components/header/Header.js
@@ -5,6 +5,9 @@ import { AiOutlineClose, AiOutlineMenu } from 'react-icons/ai';
 const Header = () => {
   const [open, setOpen] = useState(false);
 
+  const toggleMenu = () => setOpen((prevOpen) => !prevOpen);
+  const closeMenu = () => setOpen(false);
+
   return (
     <header>
       <Navbar id="navbar" bg="secondary" variant="dark" expand="md" fixed="top" collapseOnSelect>
@@ -14,18 +17,18 @@ const Header = () => {
               Dico Diaz Dussan
             </Nav.Link>
           </Navbar.Brand>
-          <Navbar.Toggle className="border-0" onClick={() => setOpen((prevOpen) => !prevOpen)}>
+          <Navbar.Toggle className="border-0" onClick={toggleMenu}>
             {open ? <AiOutlineClose /> : <AiOutlineMenu />}
           </Navbar.Toggle>
           <Navbar.Collapse>
             <Nav className="ms-auto">
-              <Nav.Link href="#portfolio" onClick={() => setOpen((prevOpen) => !prevOpen)}>
+              <Nav.Link href="#portfolio" onClick={closeMenu}>
                 Portfolio
               </Nav.Link>
-              <Nav.Link href="#about" onClick={() => setOpen((prevOpen) => !prevOpen)}>
+              <Nav.Link href="#about" onClick={closeMenu}>
                 About
               </Nav.Link>
-              <Nav.Link href="#contact" onClick={() => setOpen((prevOpen) => !prevOpen)}>
+              <Nav.Link href="#contact" onClick={closeMenu}>
                 Contact
               </Nav.Link>
             </Nav>
